feat(uploads): implement upload metadata enrichment service and types
